Move config into energex.core; add ArcticDB/Neo4j/Connector settings
